pyproject.toml
[tool.poetry]
name = "fl_with_dp_using_flwr"
version = "0.1.0"
description = ""
authors = ["shakibyzn <shakibyzn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flwr = "0.17.0"
torch = "1.10.0"
torchvision = "0.11.1"
tensorboard = "2.7.0"
torch-tb-profiler = "0.3.1"
opacus = "1.0.0"
PyYAML = "6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
